+++ src/language-markdown/print/whitespace.js
@@ -7,7 +7,7 @@ import {
   KIND_K_LETTER,
   KIND_NON_CJK,
 } from "../utilities.js";
-import { isFakeTableDelimiterRow } from "./word.js";
+import { isFakeTableDelimiterRow } from "../utilities/is-fake-table-delimiter-row.js";
 
 /**
  * @import {WordNode, WhitespaceValue, WordKind} from "../utilities.js"
